cli.commands cleanup
