+++ src/lsst/cmservice/cli/campaign.py
@@ -2,7 +2,13 @@
 
 from .. import db
 from . import options, wrappers
-from .commands import campaign_group
+from .commands import client_top
+
+
+@client_top.group(name="campaign")
+def campaign_group() -> None:
+    """Manage Campaign table"""
+
 
 # Template specialization
 # Specify the cli path to attach these commands to
