+++ src/lsst/cmservice/cli/group.py
@@ -3,7 +3,13 @@
 from .. import db
 from ..client.client import CMClient
 from . import options, wrappers
-from .commands import group_group
+from .commands import client_top
+
+
+@client_top.group(name="group")
+def group_group() -> None:
+    """Manage Group table"""
+
 
 # Template specialization
 # Specify the cli path to attach these commands to
