+++ src/lsst/cmservice/cli/job.py
@@ -3,7 +3,13 @@
 from .. import db
 from ..client.client import CMClient
 from . import options, wrappers
-from .commands import job_group
+from .commands import client_top
+
+
+@client_top.group(name="job")
+def job_group() -> None:
+    """Manage Job table"""
+
 
 # Template specialization
 # Specify the cli path to attach these commands to
