+++ src/lsst/cmservice/cli/product_set.py
@@ -2,7 +2,13 @@
 
 from .. import db
 from . import options, wrappers
-from .commands import product_set_group
+from .commands import client_top
+
+
+@client_top.group(name="product_set")
+def product_set_group() -> None:
+    """Manage ProductSet table"""
+
 
 # Template specialization
 # Specify the cli path to attach these commands to
